+++ src/prelude.rs
@@ -4,6 +4,8 @@ pub use app::{Application, ApplicationBuilder};
 pub use config::Config;
 pub use core::WithNamed;
 pub use ecs::prelude::{Builder, World};
-pub use {DataInit, GameData, GameDataBuilder};
-pub use state::{EmptyState, EmptyTrans, SimpleState, SimpleTrans, State, StateData, Trans};
+pub use state::{
+	DataInit, EmptyState, EmptyTrans, GameData, GameDataBuilder, SimpleState,
+    SimpleTrans, State, StateData, Trans
+};
 pub use events::StateEvent;